services/litr/pyproject.toml
[tool.poetry]
name = "litr"
version = "0.0.1"
description = "Test API"
authors = ["Vinodh Ravi"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "~0.14.0"
fastapi = "~0.68.0"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
Jinja2 = "~3.0.1"
SQLAlchemy = "~1.4.22"
alembic = "^1.6.5"
tenacity = "^8.0.1"
opentelemetry-api="1.5.0"
opentelemetry-sdk="1.5.0"
opentelemetry-instrumentation-fastapi="^0.24b0"
opentelemetry-exporter-otlp = "1.5.0"
opentelemetry-exporter-otlp-proto-grpc= "1.5.0"
opentelemetry-instrumentation-httpx = "^0.24b0"
rich="^10.12.0"
opentelemetry-exporter-richconsole = {path = "opentelemetry-python-contrib/exporter/opentelemetry-exporter-richconsole"}

# [tool.poetry.scripts]
# start = "app.main:start"
